Rename reloader module
